requirements.txt
datasets>=2.19.0
pytest>=8.0.0
httpx>=0.28.1
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.31.0

